+++ packages/go-kosu/abci/cli/cli.go
@@ -5,12 +5,15 @@ import (
 	"encoding/base64"
 	"fmt"
 	"os"
+	"path"
 	"strconv"
+
+	"github.com/spf13/cobra"
+	"github.com/tendermint/tendermint/config"
 
 	"github.com/ParadigmFoundation/kosu-monorepo/packages/go-kosu/abci"
 	"github.com/ParadigmFoundation/kosu-monorepo/packages/go-kosu/abci/types"
-
-	"github.com/spf13/cobra"
+	"github.com/ParadigmFoundation/kosu-monorepo/packages/go-kosu/service"
 )
 
 // CLI provides functions to build a cli client
@@ -183,4 +186,90 @@ func (cli *CLI) QueryPoster() *cobra.Command {
 			fmt.Printf("ok: < %s>\n", poster)
 		},
 	}
+}
+
+// NewInitCommand returns the init command
+func NewInitCommand(homeFlag string) *cobra.Command {
+	return &cobra.Command{
+		Use:   "init",
+		Short: "Initialize a home directory and configuration",
+		Long: "Generates a Tendermint configuration directory and key pair for a Kosu node.\n" +
+			"The default home directory is '$HOME/.kosu', which can be overridden with the '--home' flag.",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			home, err := cmd.Flags().GetString(homeFlag)
+			if err != nil {
+				return err
+			}
+
+			return abci.InitTendermint(home)
+		},
+	}
+}
+
+func getCmdCfg(cmd *cobra.Command, homeFlag string) (*config.Config, error) {
+	home, err := cmd.Flags().GetString(homeFlag)
+	if err != nil {
+		return nil, err
+	}
+	return abci.LoadConfig(home)
+}
+
+// NewNodeCommand returns the node command
+func NewNodeCommand(homeFlag string) *cobra.Command {
+	info := &cobra.Command{
+		Use:   "info",
+		Short: "Show this node's information",
+		Long:  "Displays information unique to this node",
+		RunE: func(cmd *cobra.Command, _ []string) error {
+			cfg, err := getCmdCfg(cmd, homeFlag)
+			if err != nil {
+				return err
+			}
+
+			info, err := abci.ShowNodeInfo(cfg)
+			if err != nil {
+				return err
+			}
+
+			fmt.Printf("Public Key: %s\n", info.PublicKey)
+			fmt.Printf("Node ID:    %s\n", info.NodeID)
+			fmt.Printf("Peer ID:    %s\n", info.PeerID)
+			fmt.Printf("Moniker:    %s\n", info.Moniker)
+			return nil
+		},
+	}
+
+	reset := &cobra.Command{
+		Use:   "reset",
+		Short: "Reset this node to genesis state",
+		Long:  "Reset will wipe all the data and WAL, keeping only the config and the genesis file",
+		RunE: func(cmd *cobra.Command, _ []string) error {
+			cfg, err := getCmdCfg(cmd, homeFlag)
+			if err != nil {
+				return err
+			}
+
+			abci.ResetAll(cfg, os.Stdout)
+			dbdir := path.Join(cfg.RootDir, "kosu.db")
+			return os.RemoveAll(dbdir)
+		},
+	}
+
+	cmd := &cobra.Command{
+		Use:   "node",
+		Short: "Node related operations",
+	}
+	cmd.AddCommand(info, reset)
+	return cmd
+}
+
+// NewStartCommand returns the start command
+func NewStartCommand(homeFlag string) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "start",
+		Short: "Starts the kosu node",
+		Long:  "Main entrypoint for Kosu validators and full nodes.\nPrior to use, 'kosud init' must be run.",
+	}
+	service.RegisterCommand(cmd, homeFlag)
+	return cmd
 }